backend/requirements.txt

fastapi~=0.111.0
uvicorn[standard]~=0.29.0


sqlalchemy~=2.0.30
psycopg2-binary~=2.9.9
alembic~=1.13.1


passlib[bcrypt]~=1.7.4
python-jose[cryptography]~=3.3.0
python-multipart~=0.0.9


pydantic~=2.7.1
pydantic-settings~=2.2.1
python-dotenv~=1.0.1
email_validator~=2.1.1


sendgrid~=6.11.0